requirements.txt
streamlit==0.79.0
opencv_python==4.1.2.30
imageai==2.1.5
imutils==0.5.4
Keras==2.3.1
ipython==7.22.0
numpy
Pillow==8.2.0
scikit_learn==0.24.1
tensorflow-cpu==2.3.0
h5py<3.0.0

